SSCS-8426 Updating hearing type sentence logic
