+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/writefinaldecision/uc/UcOldTemplateContent.java
@@ -1,7 +1,11 @@
 package uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.uc;
 
 import java.util.List;
+import org.apache.commons.lang3.StringUtils;
+import uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.uc.scenarios.UcTemplateComponentId;
 import uk.gov.hmcts.reform.sscs.model.docassembly.Descriptor;
+import uk.gov.hmcts.reform.sscs.model.docassembly.Paragraph;
+import uk.gov.hmcts.reform.sscs.model.docassembly.WriteFinalDecisionTemplateBody;
 
 /**
  * This is a temporary class to support the un-migrated scenarios, containing content that has initially been ported
@@ -70,6 +74,44 @@ public abstract class UcOldTemplateContent extends UcTemplateContent {
             return "The following activity and descriptor from Schedule 3 applied:";
         } else {
             return "The following activities and descriptors from Schedule 3 applied:";
+        }
+    }
+
+    public String getHearingTypeSentence(String appellantName, String bundlePage, String hearingType, boolean appellantAttended, boolean presentingOfifficerAttened) {
+        if (StringUtils.equalsIgnoreCase("paper", hearingType)) {
+            return "No party has objected to the matter being decided without a hearing. Having considered the appeal bundle to page " + bundlePage + " and the requirements of rules 2 and 27 of the Tribunal Procedure (First-tier Tribunal) (Social Entitlement Chamber) Rules 2008 the Tribunal is satisfied that it is able to decide the case in this way.";
+        } else  {
+            return getFaceToFaceTelephoneVideoHearingTypeSentence(hearingType, appellantName, bundlePage, appellantAttended, presentingOfifficerAttened);
+        }
+    }
+
+    public void addHearingType(WriteFinalDecisionTemplateBody writeFinalDecisionTemplateBody) {
+        addComponent(new Paragraph(UcTemplateComponentId.HEARING_TYPE.name(), getHearingTypeSentence(writeFinalDecisionTemplateBody.getAppellantName(), writeFinalDecisionTemplateBody.getPageNumber(),
+            writeFinalDecisionTemplateBody.getHearingType(), writeFinalDecisionTemplateBody.isAttendedHearing(), writeFinalDecisionTemplateBody.isPresentingOfficerAttended())));
+    }
+
+    public String getFaceToFaceTelephoneVideoHearingTypeSentence(String hearingType, String appellantName, String bundlePage,
+        boolean appellantAttended, boolean presentingOfifficerAttened) {
+        if (appellantAttended) {
+            if (StringUtils.equalsIgnoreCase("faceToFace", hearingType)) {
+                return "This has been an oral (face to face) hearing. "
+                    + getAppellantAttended(hearingType, appellantName, presentingOfifficerAttened, bundlePage);
+            } else {
+                return "This has been a remote hearing in the form of a " + hearingType + " hearing. "
+                    + getAppellantAttended(hearingType, appellantName, presentingOfifficerAttened, bundlePage);
+            }
+        } else {
+            if (StringUtils.equalsIgnoreCase("faceToFace", hearingType)) {
+                return appellantName + " requested an oral hearing but did not attend today. "
+                    + (presentingOfifficerAttened ? "A " : "No ") + "Presenting Officer attended on behalf of the Respondent. "
+                    + "\n"
+                    + getConsideredParagraph(bundlePage, appellantName);
+            } else {
+                return "This has been a remote hearing in the form of a " + hearingType + " hearing. " + appellantName + " did not attend the hearing today. "
+                    + (presentingOfifficerAttened ? "A" : "No") + " Presenting Officer attended on behalf of the Respondent.\n"
+                    + "\n"
+                    + getConsideredParagraph(bundlePage, appellantName);
+            }
         }
     }
 }
